Refactor PageHeader component to remove title component
