+++ apps/filecoin-site/src/app/learn/page.tsx
@@ -6,7 +6,7 @@ import { PATHS } from '@/constants/paths'
 import { Button } from '@/components/Button'
 import { Card } from '@/components/Card'
 import { Heading } from '@/components/Heading'
-import { PageHeader } from '@/components/PageHeader/PageHeader'
+import { PageHeader } from '@/components/PageHeader'
 import { PageSection } from '@/components/PageSection'
 import { SectionDivider } from '@/components/SectionDivider'
