+++ apps/filecoin-site/src/app/offer-storage/page.tsx
@@ -1,5 +1,5 @@
 import { Button } from '@/components/Button'
-import { PageHeader } from '@/components/PageHeader/PageHeader'
+import { PageHeader } from '@/components/PageHeader'
 import { PageSection } from '@/components/PageSection'
 
 export default function OfferStorage() {
