+++ apps/filecoin-site/src/app/terms-of-use/page.tsx
@@ -1,4 +1,4 @@
-import { PageHeader } from '@/components/PageHeader/PageHeader'
+import { PageHeader } from '@/components/PageHeader'
 import { PageSection } from '@/components/PageSection'
 
 export default function TermsOfUse() {
